+++ R/log_read.R
@@ -101,7 +101,12 @@ log_read <- function(
 log_read_file <- function(path, units_cpu, units_memory, units_time) {
   lines <- grep(pattern = "__AUTOMETRIC__", x = readLines(path), value = TRUE)
   lines <- gsub(".*__AUTOMETRIC__\\|(.*)\\|__AUTOMETRIC__.*", "\\1", lines)
-  out <- utils::read.table(text = lines, sep = "|", header = FALSE)
+  out <- utils::read.table(
+    text = lines,
+    sep = "|",
+    header = FALSE,
+    fill = TRUE
+  )
   colnames(out) <- c(
     "version",
     "pid",
@@ -111,8 +116,22 @@ log_read_file <- function(path, units_cpu, units_memory, units_time) {
     "core",
     "cpu",
     "resident",
+    "virtual",
+    "phase"
+  )[seq_along(out)]
+  order <- c(
+    "version",
+    "phase",
+    "pid",
+    "name",
+    "status",
+    "time",
+    "core",
+    "cpu",
+    "resident",
     "virtual"
   )
+  out <- out[, intersect(order, colnames(out))]
   out$version <- as.character(out$version)
   out$pid <- as.integer(out$pid)
   out$name <- as.character(out$name)